requirements.txt
Abg==2.3.1
pyrogram==2.0.106
pymongo
motor
requests
dnspython
tgcrypto==1.2.2
asyncio
python-dotenv
lexica-api
deep-translator
pycgi
apscheduler
Flask
uvloop
psutil
langdetect
MukeshAPI
python-dotenv